+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Imp/EpochSpec.hs
@@ -422,7 +422,7 @@ treasuryWithdrawalExpectation extraWithdrawals = do
   donateToTreasury withdrawalAmount
   committeeHotCreds <- registerInitialCommittee
   (dRepCred, _, _) <- setupSingleDRep 1_000_000
-  treasuryStart <- getsNES $ nesEsL . esChainAccountStateL . casTreasuryL
+  treasuryStart <- getsNES treasuryL
   treasuryStart `shouldBe` withdrawalAmount
   rewardAccount <- registerRewardAccount
   govPolicy <- getGovPolicy
@@ -437,7 +437,7 @@ treasuryWithdrawalExpectation extraWithdrawals = do
     getReward (raCredential rewardAccount) `shouldReturn` mempty
   passEpoch -- 2nd epoch crossing enacts all the ratified actions
   expectMissingGovActionId govActionId
-  treasuryEnd <- getsNES $ nesEsL . esChainAccountStateL . casTreasuryL
+  treasuryEnd <- getsNES treasuryL
   impAnn "Withdrawal deducted from treasury" $
     treasuryStart <-> treasuryEnd `shouldBe` withdrawalAmount
   impAnn "Withdrawal received by reward account" $
@@ -446,7 +446,7 @@ treasuryWithdrawalExpectation extraWithdrawals = do
 depositMovesToTreasuryWhenStakingAddressUnregisters :: ConwayEraImp era => ImpTestM era ()
 depositMovesToTreasuryWhenStakingAddressUnregisters = do
   disableTreasuryExpansion
-  initialTreasury <- getsNES $ nesEsL . esChainAccountStateL . casTreasuryL
+  initialTreasury <- getsNES treasuryL
   modifyPParams $ \pp ->
     pp
       & ppGovActionLifetimeL .~ EpochInterval 8