+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.18.0.0
 
+* Add `CanGetChainAccountState` and `CanSetChainAccountState`
+* Add `treasuryL` and `reservesL`
 * Add `casTreasuryL` and `casReservesL`
 * Rename `AccountState` to `ChainAccountState`. Rename accessor fields `asTreasury` and `asReserves` to `casTreasury` and `casReserves` respectively.
 * Added `sizeTxForFeeCalculation` to `EraTx` with a default implementation